RT-37: Исправлены методы для CORS, проверки тестовых данных, хендлеры вынесены в отдедьный файл.
