+++ internal/pkg/user/userauth_test.go
@@ -20,24 +20,29 @@ func TestHandlerSignup(t *testing.T) {
 	tests := []TestCase{
 		{"", `{"error":"bad form"}`, http.StatusBadRequest},
 		{`{"bad_field":bad_value}`, `{"error":"bad form"}`, http.StatusBadRequest},
-		{`{"username":"a","email":"ya@mail.ru","password":"pass","repeated_password":"not_pass"}`,
+		{`{"username":"a","email":"ya@mail.ru","password":"pass","confirm_password":"not_pass"}`,
 			`{"error":"Passwords do not match"}`, http.StatusBadRequest},
 		{"{}", `{"error":"Empty fields in form"}`, http.StatusBadRequest},
-		{`{"username":"good_user","email":"gmail@mail.ru","password":"pass","repeated_password":"pass"}`,
+		{`{"username":"good_user","email":"gmail@mail.ru","password":"pass","confirm_password":"pass"}`,
 			`{"id":1,"email":"gmail@mail.ru","username":"good_user"}`, http.StatusOK},
-		{`{"username":"good_user","email":"gmail@mail.ru","password":"pass","repeated_password":"pass"}`,
+		{`{"username":"good_user","email":"gmail@mail.ru","password":"pass","confirm_password":"pass"}`,
 			`{"error":"Wrong username or password"}`, http.StatusBadRequest},
 	}
 
 	api := &Handler{}
 	for i, test := range tests {
+		test.outJSON += "\n"
 		fmt.Println("TestSignup", i)
 		body := bytes.NewReader([]byte(test.inJSON))
 		r := httptest.NewRequest("POST", "/users/signup", body)
 		w := httptest.NewRecorder()
 		api.Signup(w, r)
-		require.Equal(t, test.status, w.Code)
-		require.Equal(t, test.outJSON+"\n", w.Body.String())
+		actual := TestCase{
+			inJSON:  test.inJSON,
+			outJSON: w.Body.String(),
+			status:  w.Code,
+		}
+		require.Equal(t, test, actual)
 	}
 }
 
@@ -50,13 +55,18 @@ func TestHandlerLogin(t *testing.T) {
 
 	api := &Handler{}
 	for i, test := range testForm {
+		test.outJSON += "\n"
 		fmt.Println("TestLogin (form parsing)", i)
 		body := bytes.NewReader([]byte(test.inJSON))
 		r := httptest.NewRequest("POST", "/users/login", body)
 		w := httptest.NewRecorder()
 		api.Login(w, r)
-		require.Equal(t, test.status, w.Code)
-		require.Equal(t, test.outJSON+"\n", w.Body.String())
+		actual := TestCase{
+			inJSON:  test.inJSON,
+			outJSON: w.Body.String(),
+			status:  w.Code,
+		}
+		require.Equal(t, test, actual)
 	}
 
 	testData := []TestCase{
@@ -67,17 +77,22 @@ func TestHandlerLogin(t *testing.T) {
 		{`{"email":"gmail@mail.ru","password":"pass"}`,
 			`{"id":1,"email":"gmail@mail.ru","username":"good_user"}`, http.StatusOK},
 	}
-	testUser := `{"username":"good_user","email":"gmail@mail.ru","password":"pass","repeated_password":"pass"}`
+	testUser := `{"username":"good_user","email":"gmail@mail.ru","password":"pass","confirm_password":"pass"}`
 	api.Signup(httptest.NewRecorder(),
 		httptest.NewRequest("POST", "/users/signup", bytes.NewReader([]byte(testUser))))
 
 	for i, test := range testData {
+		test.outJSON += "\n"
 		fmt.Println("TestLogin (data checking)", i)
 		body := bytes.NewReader([]byte(test.inJSON))
 		r := httptest.NewRequest("POST", "/users/login", body)
 		w := httptest.NewRecorder()
 		api.Login(w, r)
-		require.Equal(t, test.status, w.Code)
-		require.Equal(t, test.outJSON+"\n", w.Body.String())
+		actual := TestCase{
+			inJSON:  test.inJSON,
+			outJSON: w.Body.String(),
+			status:  w.Code,
+		}
+		require.Equal(t, test, actual)
 	}
 }
